+++ customer/checkout.php
@@ -1,122 +1,5 @@
-<!-- templatemo 385 floral shop -->
-<!-- 
-Floral Shop Template 
-http://www.templatemo.com/preview/templatemo_385_floral_shop 
--->
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml">
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-<title>Check Out - Floral Template</title>
-<meta name="keywords" content="" />
-<meta name="description" content="" />
-<link href="templatemo_style.css" rel="stylesheet" type="text/css" />
-
-<link rel="stylesheet" href="css/orman.css" type="text/css" media="screen" />
-<link rel="stylesheet" href="css/nivo-slider.css" type="text/css" media="screen" />
-
-<link rel="stylesheet" type="text/css" href="css/ddsmoothmenu.css" />
-
-<script type="text/javascript" src="js/jquery.min.js"></script>
-<script type="text/javascript" src="js/ddsmoothmenu.js">
-
-/***********************************************
-* Smooth Navigational Menu- (c) Dynamic Drive DHTML code library (www.dynamicdrive.com)
-* This notice MUST stay intact for legal use
-* Visit Dynamic Drive at http://www.dynamicdrive.com/ for full source code
-***********************************************/
-
-</script>
-
-<script type="text/javascript">
-
-ddsmoothmenu.init({
-	mainmenuid: "templatemo_menu", //menu DIV id
-	orientation: 'h', //Horizontal or vertical menu: Set to "h" or "v"
-	classname: 'ddsmoothmenu', //class added to menu's outer DIV
-	//customtheme: ["#1c5a80", "#18374a"],
-	contentsource: "markup" //"markup" or ["container_id", "path_to_menu_file"]
-})
-
-function clearText(field)
-{
-    if (field.defaultValue == field.value) field.value = '';
-    else if (field.value == '') field.value = field.defaultValue;
-}
-
-</script>
-
-</head>
-
-<body>
-
-<div id="templatemo_wrapper_sp">
-<div id="templatemo_header_wsp">
-	<div id="templatemo_header" class="header_subpage">
-    	<div id="site_title"><a href="#">Floral Shop</a></div>
-        <div id="templatemo_menu" class="ddsmoothmenu">
-            <ul>
-                <li><a href="index.html">Home</a></li>
-                <li><a href="about.html">About</a></li>
-                <li><a href="products.html">Products</a>
-                    <ul>
-                        <li><a href="#subpage1">Sub Page One</a></li>
-                        <li><a href="#subpage2">Sub Page Two</a></li>
-                        <li><a href="#subpage3">Sub Page Three</a></li>
-                        <li><a href="#subpage4">Sub Page Four</a></li>
-                        <li><a href="#subpage5">Sub Page Five</a></li>
-                  	</ul>
-                </li>
-                <li><a href="checkout.html" class="selected">Checkout</a></li>
-                <li><a href="contact.html">Contact</a></li>
-                <li><a href="faqs.html">FAQs</a></li>
-            </ul>
-            <div id="templatemo_search">
-                <form action="#" method="get">
-                  	<input type="text" value="Site Search" name="keyword" id="keyword" title="keyword" 
-                  			onfocus="clearText(this)" onblur="clearText(this)" class="txt_field" />
-                  	<input type="submit" name="Search" value="" alt="Search" id="searchbutton" title="Search" class="sub_btn" />
-                </form>
-            </div>
-            <br style="clear: left" />
-        </div> <!-- end of templatemo_menu -->
-
-    </div> <!-- END of header -->
-</div> <!-- END of header wrapper -->
-<div id="templatemo_main_wrapper">
-<div id="templatemo_main">
-	<div id="sidebar" class="left">
-    	<div class="sidebar_box"><span class="bottom"></span>
-            <h3>Categories</h3>   
-            <div class="content"> 
-                <ul class="sidebar_list">
-                    <li><a href="#">Nulla odio ipsum</a></li>
-                    <li><a href="#">Suspendisse posuere</a></li>
-                    <li><a href="#">Aliquam euismod</a></li>
-                    <li><a href="#">Curabitur ac mauris</a></li>
-                    <li><a href="#">Mauris nulla tortor</a></li>
-                    <li><a href="#">Nullam ultrices</a></li>
-                    <li><a href="#">Vivamus scelerisque</a></li>
-                    <li><a href="#">Suspendisse posuere</a></li>
-                    <li><a href="#">Quisque vel justo</a></li>
-                </ul>
-            </div>
-        </div>
-        <div class="sidebar_box"><span class="bottom"></span>
-            <h3>Weekly Special</h3>   
-            <div class="content special"> 
-                <img src="images/templatemo_image_01.jpg" alt="Flowers" />
-                <a href="#">Citrus Burst Bouquet</a>
-                <p>
-                	Price:
-                    <span class="price normal_price">$160</span>&nbsp;&nbsp;
-                    <span class="price special_price">$100</span>
-                </p>
-            </div>
-        </div>
-    </div>
-    
-    <div id="content" class="right">
+<?php require ('forAll.php');?>
+<div id="content" class="right">
     	<h2>Checkout</h2>
         
 		<h3>BILLING DETAILS</h3>
